+++ trunk/platforms/osx/logjoint.mac/ui/Postprocessors/StateInspectorWindow/PropertiesViewDataSource.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using MonoMac.AppKit;
+using AppKit;
 using System.Collections.Generic;
 
 namespace LogJoint.UI.Postprocessing.StateInspector
@@ -8,7 +8,7 @@ namespace LogJoint.UI.Postprocessing.StateInspector
 	{
 		public IList<KeyValuePair<string, object>> data = new KeyValuePair<string, object>[0];
 
-		public override int GetRowCount (NSTableView tableView)
+		public override nint GetRowCount (NSTableView tableView)
 		{
 			return data.Count;
 		}
